Crate tokio_socks

Modules

tcp

Structs

ProxyAddrsStream

Enums

Error
Error type of tokio-socks

TargetAddr
A SOCKS connection target.

Traits

IntoTargetAddr
A trait for objects that can be converted to TargetAddr.

ToProxyAddrs
A trait for objects which can be converted or resolved to one or more SocketAddr values, which are going to be connected as the the proxy server.

Type Definitions

Result