Struct tokio_socks::tcp::Socks5Stream

pub struct Socks5Stream { /* fields omitted */ }

A SOCKS5 client.

For convenience, it can be dereferenced to tokio_tcp::TcpStream.

Methods

impl Socks5Stream

pub async fn connect<'t, P, T>(proxy: P, target: T) -> Result<Self> where
P: ToProxyAddrs,
T: IntoTargetAddr<'t>,

Connects to a target server through a SOCKS5 proxy.

Error

It propagates the error that occurs in the conversion from T to TargetAddr.

pub async fn connect_with_password<'a, 't, P, T>(
proxy: P,
target: T,
username: &'a str,
password: &'a str
) -> Result<Self> where
P: ToProxyAddrs,
T: IntoTargetAddr<'t>,

Connects to a target server through a SOCKS5 proxy using given username and password.

Error

It propagates the error that occurs in the conversion from T to TargetAddr.

pub fn into_inner(self) -> TcpStream

Consumes the Socks5Stream, returning the inner tokio_tcp::TcpStream.

pub fn target_addr(&self) -> TargetAddr

Returns the target address that the proxy server connects to.