Trait tokio_socks::ToProxyAddrs

pub trait ToProxyAddrs {
    type Output: Stream<Item = Result<SocketAddr>> + Unpin;
    fn to_proxy_addrs(&self) -> Self::Output;
}

A trait for objects which can be converted or resolved to one or more SocketAddr values, which are going to be connected as the the proxy server.

This trait is similar to std::net::ToSocketAddrs but allows asynchronous name resolution.

Associated Types

type Output: Stream<Item = Result<SocketAddr>> + Unpin

Required methods

fn to_proxy_addrs(&self) -> Self::Output